flashloans-lab/test_rust.py
#!/usr/bin/env python3
"""
Main entry point for the Flash Loan Arbitrage Bot
Run this to start monitoring for arbitrage opportunities
"""

import asyncio
import sys
import os
from datetime import datetime

# Add the project root to path
sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))

async def main():
    print("""
    ╔══════════════════════════════════════════════════════════════╗
    ║           FLASH LOAN ARBITRAGE BOT - RESEARCH MODE          ║
    ╠══════════════════════════════════════════════════════════════╣
    ║  ⚡ High-Performance DeFi Arbitrage Engine                   ║
    ║  🦀 Powered by Rust Core Math Library                       ║
    ║  📊 100 Strategies | 10 Families | Multi-Chain              ║
    ╚══════════════════════════════════════════════════════════════╝
    """)
    
    print(f"[{datetime.now().strftime('%H:%M:%S')}] Initializing components...")
    
    # Import components
    from engine.config import Config
    from engine.strategies.registry import StrategyRegistry
    from engine.execsim.simulator import Simulator
    from engine.execsim.portfolio import PortfolioManager, RiskLimits
    from engine.pricing.route_search import RouteSearcher
    from engine.datasources.dex_index import DEXIndexer, PoolInfo
    from engine.datasources.evm_rpc import EVMRPCClient
    from engine.datasources.gas_tracker import GasTracker
    
    # Test Rust module
    try:
        import flashloans_core
        print(f"[{datetime.now().strftime('%H:%M:%S')}] ✅ Rust core loaded successfully")
    except ImportError:
        print(f"[{datetime.now().strftime('%H:%M:%S')}] ⚠️  Rust core not found, using Python fallback")
    
    # Load configuration
    print(f"[{datetime.now().strftime('%H:%M:%S')}] Loading configuration...")
    config = Config.load("configs")
    
    # Initialize strategy registry
    print(f"[{datetime.now().strftime('%H:%M:%S')}] Loading strategies...")
    registry = StrategyRegistry()
    strategies = registry.load_strategies_from_config("configs/strategies.yaml")
    enabled_strategies = registry.get_enabled_strategies()
    print(f"[{datetime.now().strftime('%H:%M:%S')}] 📊 Loaded {len(strategies)} strategies ({len(enabled_strategies)} enabled)")
    
    # Initialize portfolio manager
    risk_limits = RiskLimits(
        max_drawdown_pct=5.0,
        max_position_size=1_000_000,
        max_daily_trades=100,
        kelly_fraction=0.25
    )
    portfolio = PortfolioManager(risk_limits)
    
    # Initialize execution simulator
    simulator = ExecutionSimulator(config)
    
    # Simulated monitoring loop
    print(f"\n[{datetime.now().strftime('%H:%M:%S')}] Starting monitoring loop...")
    print(f"[{datetime.now().strftime('%H:%M:%S')}] Mode: SIMULATION (no real transactions)")
    
    # Demo: Create mock market data
    mock_pools = [
        {
            'address': '0x1234...',
            'pool_type': 'uniswap_v2',
            'token0': '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2',  # WETH
            'token1': '0xA0b86991c444844b40458E39Ce1e19b0c39A5d95',  # USDC
            'reserve0': 10000000000000000000000,  # 10,000 WETH
            'reserve1': 20000000000000000000000,  # 20,000 USDC
            'fee': 30
        },
        {
            'address': '0x5678...',
            'pool_type': 'uniswap_v2',
            'token0': '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2',  # WETH
            'token1': '0xA0b86991c444844b40458E39Ce1e19b0c39A5d95',  # USDC
            'reserve0': 10100000000000000000000,  # 10,100 WETH (slightly different)
            'reserve1': 19800000000000000000000,  # 19,800 USDC
            'fee': 30
        }
    ]
    
    market_data = {
        'pools': {p['address']: p for p in mock_pools},
        'gas_price': 20_000_000_000,  # 20 gwei
        'block': {'number': 18000000, 'timestamp': int(datetime.now().timestamp())}
    }
    
    opportunities_found = 0
    total_profit = 0
    
    for i in range(5):  # Run 5 simulation cycles
        print(f"\n[{datetime.now().strftime('%H:%M:%S')}] 🔍 Scanning cycle {i+1}/5...")
        
        # Check each strategy
        for name, strategy in enabled_strategies.items():
            result = await strategy.detect(market_data)
            
            if result.opportunity_found:
                opportunities_found += 1
                print(f"[{datetime.now().strftime('%H:%M:%S')}] 💡 Opportunity found: {name}")
                print(f"  - Edge: {result.edge_bps:.2f} bps")
                print(f"  - Confidence: {result.confidence:.1%}")
                print(f"  - Risk score: {result.risk_score:.2f}")
                
                if result.candidate_trade:
                    # Simulate execution
                    gas_estimate = {'max_fee': 30_000_000_000}  # 30 gwei
                    sim_result = await simulator.simulate_trade(
                        result.candidate_trade,
                        gas_estimate,
                        market_data['block']['number']
                    )
                    
                    if sim_result.success:
                        print(f"  ✅ Simulation successful!")
                        print(f"  - Expected profit: ${sim_result.net_pnl / 1e18:.2f}")
                        print(f"  - Gas cost: ${sim_result.gas_cost / 1e18:.2f}")
                        total_profit += sim_result.net_pnl
                        
                        # Record in portfolio
                        portfolio.record_trade(sim_result)
                    else:
                        print(f"  ❌ Simulation failed (revert probability: {sim_result.revert_probability:.1%})")
        
        # Brief pause between cycles
        await asyncio.sleep(2)
    
    # Print summary
    print(f"\n{'='*60}")
    print(f"SIMULATION SUMMARY")
    print(f"{'='*60}")
    print(f"Opportunities found: {opportunities_found}")
    print(f"Total simulated profit: ${total_profit / 1e18:.2f}")
    
    portfolio_summary = portfolio.get_portfolio_summary()
    print(f"\nPortfolio Statistics:")
    print(f"  - Total trades: {portfolio_summary['total_trades']}")
    print(f"  - Win rate: {portfolio_summary['global_win_rate']:.1%}")
    print(f"  - Global P&L: ${portfolio_summary['global_pnl'] / 1e18:.2f}")
    
    print(f"\n[{datetime.now().strftime('%H:%M:%S')}] Simulation complete!")
    print("\n⚠️  This was a SIMULATION. No real transactions were executed.")
    print("To run in production, you would need:")
    print("  1. Real RPC endpoints (Alchemy/Infura)")
    print("  2. Private keys for transaction signing")
    print("  3. Sufficient ETH for gas")
    print("  4. Proper risk management settings")

if __name__ == "__main__":
    asyncio.run(main())