flashloans-lab/requirements.txt
# Core blockchain libraries
web3==6.11.0
eth-abi==4.2.1

# Data and validation
pydantic==2.4.0
numpy==1.26.4  # Updated for Python 3.13 compatibility
pandas==2.1.4  # Updated for compatibility

# Async and networking
websockets==12.0
aiohttp==3.9.1
uvicorn==0.24.0
fastapi==0.104.1

# Monitoring and optimization
prometheus-client==0.19.0
optuna==3.4.0
scikit-learn==1.3.2
scipy==1.11.4

# Data handling
pyarrow==14.0.1
tenacity==8.2.3

# Development tools
mypy==1.7.1
ruff==0.1.6

# CLI and config
click==8.1.7
pyyaml==6.0.1

# Additional dependencies for compatibility
setuptools==69.0.2
wheel==0.42.0